[pending]: Products FS persistence

+++ challenge-07/src/dao/productsMongoDAO.js
@@ -1,4 +1,6 @@
+import mongoose from 'mongoose';
 import { productsModel } from "./models/products.model.js";
+
 
 export class ProductsMongoDAO{
     constructor(){}
@@ -30,6 +32,20 @@ export class ProductsMongoDAO{
     }
     
     async delete(id){
+        if(!mongoose.Types.ObjectId.isValid(id)){
+            // Invalid ObjectId Middleware
+            throw new Error('El pid ingresado tiene un formato invalido');
+        }
+        
+        let products = await this.get();
+        let prodId = products.filter(product => product._id.equals(new mongoose.Types.ObjectId(id)));
+        
+        if(prodId.length === 0){
+            // Invalid Pid Middleware
+            throw new Error(`El producto con PID ${id} no existe`); // Caso en el que se cumple http://localhost:8080/api/products/34123123
+        }
+        
         return await productsModel.deleteOne({_id:id});
     }
 }
+
